+++ CHANGELOG.md
@@ -2,8 +2,9 @@ Percona Monitoring and Management (PMM) Client
 
 v1.0.7 unreleased 2016-11-28
 
-* Automatically generate self-signed SSL certificate to protect metric services with HTTPS/TLS by default (applies to newerly added services, see "check-network" output).
-* Enable http basic auth for metric services automatically when defined on PMM server and configured on client to achieve client-side protection.
+* Automatically generate self-signed SSL certificate to protect metric services with HTTPS/TLS by default (requires re-adding services, see "check-network" output).
+* Enable http basic auth for metric services when defined on PMM server and configured on client to achieve client-side protection
+  (requires re-adding services, see "check-network" output).
 * Added --bind-address flag to support running PMM server and client on the different networks.
   By default, this address is the same as client one. When running PMM on different networks, --client-address should be set to remote (public) address
   and --bind-address to local (private) address. This also assumes you configure NAT and port forwarding between those addresses.